 requirements.txt
requests
beautifulsoup4
streamlit-pydantic
pydantic==1.10.14

